requirements.txt
alembic==1.4.0
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==19.10.0
idna==2.9
injector==0.18.3
itsdangerous==1.1.0
Jinja2==2.11.1
jsonrpclib==0.1.7
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==2.20.5
marshmallow-sqlalchemy==0.18.0
passlib==1.7.2
psycopg2==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
Werkzeug==1.0.0
